+++ Engine_SOURCE/yaActionScript.cpp
@@ -34,20 +34,8 @@ namespace ya
 		, mSpeed(100.0f)
 		, mDirection(Vector3::Zero)
 		, mRotateDirection(Vector3::Zero)
-		, mGroundDistance(0.f)
-		, mGroundNormal(Vector3::Zero)
-		, mGroundSlopeAngle(0.f)
-		, mForwardSlopeAngle(0.f)
-		, mGroundCross(0.f)
-		, mbMoving(false)
-		, mbRunning(false)
-		, mbJumping(false)
-		, mbGrounded(false)
-		, mbForwardBlocked(false)
 		, mJumpTimer(0.0f)
 		, mJumpForce(0.0f)
-		, mJumpEvent(nullptr)
-		, mGroundEvent(nullptr)
 	{
 	}
 
@@ -84,8 +72,6 @@ namespace ya
 	{
 		assert(GetOwner() != nullptr);
 
-		CheckGround();
-
 		if (mJumpTimer > 0.0f)
 		{
  			mJumpTimer -= Time::DeltaTime();
@@ -109,31 +95,10 @@ namespace ya
 
 	void ActionScript::OnCollisionEnter(Collider2D* collider)
 	{
-		//GameObject* colObj = collider->GetOwner();
-		//Transform* colTransform = colObj->GetComponent<Transform>();
-
-		//GameObject* obj = GetOwner();
-		//Transform* objTransform = obj->GetComponent<Transform>();
-
-		//// 벽 충돌
-		//if (nullptr != colObj->GetScript<WallScript>())
-		//{
-		//}
 	}
 
 	void ActionScript::OnCollisionStay(Collider2D* collider)
 	{
-		//GameObject* colObj = collider->GetOwner();
-		//Transform* colTransform = colObj->GetComponent<Transform>();
-
-		//GameObject* obj = GetOwner();
-		//Transform* objTransform = obj->GetComponent<Transform>();
-
-		//// 벽 충돌
-		//if (nullptr != colObj->GetScript<WallScript>())
-		//{
-		//	Rigidbody* objRigidbody = obj->GetComponent<Rigidbody>();
-		//}
 	}
 
 	void ActionScript::OnCollisionExit(Collider2D* collider)
@@ -154,6 +119,8 @@ namespace ya
 		}
 
 		mDirection = dir;
+
+
 
 		if (speed > 0.0f)
 			mRigidbody->AddForce(speed * mDirection);
@@ -213,7 +180,7 @@ namespace ya
   			mJumpForce = defaultJumpForce;
 		}
 
-		if (mbGrounded)
+		if (mRigidbody->IsGrounded())
 		{
 			mJumpTimer = 0.1f;
 		}
@@ -248,10 +215,14 @@ namespace ya
 		Vector3 position = mTransform->GetPosition();
 		Vector3 scale = mTransform->GetScale();
 		Vector3 colScale = mCollider->GetSize();
-		Vector3 velocity = movement;
-		Vector3 dir = velocity;
+		Vector3 velocity = movement * Time::DeltaTime();
+		Vector3 dir = movement;
 		dir.Normalize();
+
 		float velocityLength = velocity.Length();
+		float positionLength = position.Length();
+
+		position += dir * 5.f;
 
 		colScale *= scale;
 
@@ -275,140 +246,131 @@ namespace ya
 
 		for (int i = 0; i < 3; ++i)
 		{
-			if (velocity.Length() <= ForwardHit[i].length && ForwardHit[i].isHit)
-			{
-				//mRigidbody->SetVelocity(Vector3::Zero);
-				mbForwardBlocked = true;
-			}
-			else
-				mbForwardBlocked = false;
-		}
+			if (velocity.Length() >= ForwardHit[i].length && ForwardHit[i].isHit)
+				return true;
 
-		return false;
-	}
-
-	// 땅, 경사로 체크. 일단 가운데 레이만 사용함..
-	void ActionScript::CheckGround()
-	{
-		Vector3 position = mTransform->GetPosition();
-		Vector3 objPos = mCollider->GetCenter();
-		Vector3 Scale = mTransform->GetScale();
-		Vector3 objScale = mCollider->GetSize();
-
-		position += objPos;
-		objScale *= Scale;
-
-		// 지형체크용 레이의 시작점은 포지션의 맨아래에서 시작
-		Vector3 rayPosition = position;//mTransform->GetPosition();
-		rayPosition.y -= objScale.y / 2.f;
-
-		// z의 크기 절반 만큼의 크기를 가진 forward 벡터와 위에서 설정한 위치 벡터를 더하여
-		// forward 방향을 가진 z크기의 절반 값을 가진 벡터를 구한다
-		Vector3 forward = mTransform->Forward();
-		forward = forward * objScale.z / 2.f;
-		Vector3 forwardPos = rayPosition + forward;
-
-		// 순서대로 북서, 북동, 남서, 남동
-		Vector3 nw = forwardPos;
-		nw.x -= objScale.x / 2.f;
-
-		Vector3 ne = forwardPos;
-		ne.x += objScale.x / 2.f;
-
-		Vector3 sw = forwardPos;
-		sw.x -= objScale.x / 2.f;
-		sw.z -= objScale.z;
-
-		Vector3 se = forwardPos;
-		se.x += objScale.x / 2.f;
-		se.z -= objScale.z;
-
-		std::vector<eLayerType> layers = {};
-		layers.push_back(eLayerType::Ground);
-
-		Vector3 direction = -(mTransform->Up());
-
-		// 지형 체크용
-		RayHit CheckHit[4] = {};
-		CheckHit[0] = CollisionManager::RayCast(GetOwner(), nw, direction, layers);
-		CheckHit[1] = CollisionManager::RayCast(GetOwner(), ne, direction, layers);
-		CheckHit[2] = CollisionManager::RayCast(GetOwner(), sw, direction, layers);
-		CheckHit[3] = CollisionManager::RayCast(GetOwner(), se, direction, layers);
-
-		// 지형 보정용
-		RayHit CorrectionHit = CollisionManager::RayCast(GetOwner(), position, direction, layers);
-	
-		// 땅에 붙었는지 체크하기 위한 플래그
-		bool bGroundFlag = false;
-
-		for (int i = 0; i < 4; ++i)
-		{
-			if (CheckHit[i].isHit)
-			{
-				Transform* hitTransform = CheckHit[i].hitObj->GetComponent<Transform>();
-
-				mGroundNormal = hitTransform->Up();
-
-				float groundRadian = mGroundNormal.Dot(Vector3::Up);
-				groundRadian = acos(groundRadian);
-
-				float forwardRadian = mGroundNormal.Dot(mDirection);
-				forwardRadian = acos(forwardRadian);
-
-				mGroundSlopeAngle = groundRadian;
-				mForwardSlopeAngle = forwardRadian - 1.5708f;
-
-				// 밑바닥을 기준으로 구한 거리
-				float CorrectionLength = CorrectionHit.length - objScale.y / 2.f;
-
-				if (CorrectionLength < 0.001f)
-				{ // 바닥으로 부터 거리가 0.001f 미만일때
-					if (!mbGrounded && mJumpTimer < 0.0f)
-					{	// 착지하는지 체크
-						// ground가 아닌 상태(공중일 때)에서 jumpTimer가 종료 됐을 때(jump가 끝났을 때) jump->ground로 상태가 변화한 것으로 판단한다.
-						// jumpTimer가 0.0f 보다 크면 jump를 시작한 것이므로 jump 상태를 false로 바꾸지 않는다.
-						SetJumping(false);
-
-						if (mGroundEvent != nullptr)
-							mGroundEvent();
-					}
-
-					mbGrounded = true;
-					bGroundFlag |= true;
-
-					// 물체의 y 크기의 절반과 가운데에서 쏜 레이의 길이를 뺐을때
-					// 수치가 0보다 크면 물체는 땅을 뚫었다고 판단.
-					float overPos = objScale.y / 2.f - CorrectionHit.length;
-					Vector3 correctionPos = mTransform->GetPosition();
-
-					// 땅을 뚫을때
-					if (0.f < overPos)
-					{
-						correctionPos.y += overPos;
-						mTransform->SetPosition(correctionPos);
-					}
-				}
-				break;
-			}
-		}
-
-		if (!bGroundFlag)
-		{
-			if (mbGrounded)
-				if(mJumpEvent != nullptr)
-					mJumpEvent();
-
-			mbGrounded = false;
-		}
-
-		mGroundCross = mGroundNormal.Cross(Vector3::Up);
-
-		if (Vector3::Zero != mGroundCross)
-		{
-			Vector3 axis = mGroundCross;
-
-			Matrix mat = Matrix::CreateFromAxisAngle(axis, -mGroundSlopeAngle);
-			mRigidbody->SetRotateDirection(mat);
+			return false;
 		}
 	}
+
+	//// 땅, 경사로 체크. 일단 가운데 레이만 사용함..
+	//void ActionScript::CheckGround()
+	//{
+	//	Vector3 position = mTransform->GetPosition();
+	//	Vector3 objPos = mCollider->GetCenter();
+	//	Vector3 Scale = mTransform->GetScale();
+	//	Vector3 objScale = mCollider->GetSize();
+
+	//	position += objPos;
+	//	objScale *= Scale;
+
+	//	// 지형체크용 레이의 시작점은 포지션의 맨아래에서 시작
+	//	Vector3 rayPosition = position;
+	//	rayPosition.y -= objScale.y / 2.f;
+
+	//	// z의 크기 절반 만큼의 크기를 가진 forward 벡터와 위에서 설정한 위치 벡터를 더하여
+	//	// forward 방향을 가진 z크기의 절반 값을 가진 벡터를 구한다
+	//	Vector3 forward = mTransform->Forward();
+	//	forward = forward * objScale.z / 2.f;
+	//	Vector3 forwardPos = rayPosition + forward;
+
+	//	// 순서대로 북서, 북동, 남서, 남동
+	//	//Vector3 nw = forwardPos;
+	//	//nw.x -= objScale.x / 2.f;
+
+	//	//Vector3 ne = forwardPos;
+	//	//ne.x += objScale.x / 2.f;
+
+	//	//Vector3 sw = forwardPos;
+	//	//sw.x -= objScale.x / 2.f;
+	//	//sw.z -= objScale.z;
+
+	//	//Vector3 se = forwardPos;
+	//	//se.x += objScale.x / 2.f;
+	//	//se.z -= objScale.z;
+
+	//	std::vector<eLayerType> layers = {};
+	//	layers.push_back(eLayerType::Ground);
+
+	//	Vector3 direction = -(mTransform->Up());
+
+	//	// 지형 체크용
+	//	//RayHit CheckHit[4] = {};
+	//	//CheckHit[0] = CollisionManager::RayCast(GetOwner(), nw, direction, layers);
+	//	//CheckHit[1] = CollisionManager::RayCast(GetOwner(), ne, direction, layers);
+	//	//CheckHit[2] = CollisionManager::RayCast(GetOwner(), sw, direction, layers);
+	//	//CheckHit[3] = CollisionManager::RayCast(GetOwner(), se, direction, layers);
+
+	//	// 지형 보정용
+	//	RayHit CorrectionHit = CollisionManager::RayCast(GetOwner(), position, direction, layers);
+
+	//	bool bGroundFlag = false;
+	//
+	//	if (CorrectionHit.isHit)
+	//	{
+	//		// 땅에 붙었는지 체크하기 위한 플래그
+	//		Transform* hitTransform = CorrectionHit.hitObj->GetComponent<Transform>();
+
+	//		mGroundNormal = hitTransform->Up();
+
+	//		float groundRadian = mGroundNormal.Dot(Vector3::Up);
+	//		groundRadian = acos(groundRadian);
+
+	//		float forwardRadian = mGroundNormal.Dot(mDirection);
+	//		forwardRadian = acos(forwardRadian);
+
+	//		mGroundSlopeAngle = groundRadian;
+	//		mForwardSlopeAngle = forwardRadian - 1.5708f;
+
+	//		// 밑바닥을 기준으로 구한 거리
+	//		float CorrectionLength = CorrectionHit.length - objScale.y / 2.f;
+
+	//		if (CorrectionLength < 0.001f)
+	//		{ // 바닥으로 부터 거리가 0.001f 미만일때
+	//			if (!mbGrounded && mJumpTimer < 0.0f)
+	//			{	// 착지하는지 체크
+	//				// ground가 아닌 상태(공중일 때)에서 jumpTimer가 종료 됐을 때(jump가 끝났을 때) jump->ground로 상태가 변화한 것으로 판단한다.
+	//				// jumpTimer가 0.0f 보다 크면 jump를 시작한 것이므로 jump 상태를 false로 바꾸지 않는다.
+	//				SetJumping(false);
+
+	//				if (mGroundEvent != nullptr)
+	//					mGroundEvent();
+	//			}
+
+	//			mbGrounded = true;
+	//			bGroundFlag |= true;
+
+	//			// 물체의 y 크기의 절반과 가운데에서 쏜 레이의 길이를 뺐을때
+	//			// 수치가 0보다 크면 물체는 땅을 뚫었다고 판단.
+	//			float overPos = objScale.y / 2.f - CorrectionHit.length;
+	//			Vector3 correctionPos = mTransform->GetPosition();
+
+	//			// 땅을 뚫을때
+	//			if (0.f < overPos)
+	//			{
+	//				correctionPos.y += overPos;
+	//				mTransform->SetPosition(correctionPos);
+	//			}
+	//		}
+	//	}
+
+	//	if (!bGroundFlag)
+	//	{
+	//		if (mbGrounded)
+	//			if(mJumpEvent != nullptr)
+	//				mJumpEvent();
+
+	//		mbGrounded = false;
+	//	}
+
+	//	mGroundCross = mGroundNormal.Cross(Vector3::Up);
+
+	//	if (Vector3::Zero != mGroundCross)
+	//	{
+	//		Vector3 axis = mGroundCross;
+
+	//		Matrix mat = Matrix::CreateFromAxisAngle(axis, -mGroundSlopeAngle);
+	//		mRigidbody->SetRotateDirection(mat);
+	//	}
+	//}
 }
