+++ Engine_SOURCE/yaActionScript.h
@@ -26,14 +26,6 @@ namespace ya
 		void SetDirection(const Vector3 dir) { mDirection = dir; }
 		Vector3 GetDirection() const { return mDirection; }
 
-		void SetJumping(bool jumping) { mbJumping = jumping; }
-		bool IsJumping() { return mbJumping; }
-		void SetGrounded(bool grounded) { mbGrounded = grounded; }
-		bool IsGrounded() { return mbGrounded; }
-
-		std::function<void()>& GetJumpEvent() { return mJumpEvent; }
-		std::function<void()>& GetGroundEvent() { return mGroundEvent; }
-
 	public:
 		void Move(const Vector3 dir, float force = -1.0f);
 		void Rotate(const Vector3 dir, float speed = -1.0f);
@@ -43,7 +35,6 @@ namespace ya
 		void Parrying();
 
 		bool ForwardCheck(Vector3 movement);
-		void CheckGround();
 
 	protected:
 		GameObject* mTarget;
@@ -56,24 +47,6 @@ namespace ya
 		float mSpeed;
 		Vector3 mDirection;
 		Vector3 mRotateDirection;
-
-		Vector3 mGroundNormal;
-		Vector3 mGroundCross;
-		float mGroundDistance;
-		float mGroundSlopeAngle;
-		float mForwardSlopeAngle;
-
-		bool mbMoving;
-		bool mbRunning;
-		bool mbJumping;
-		bool mbGrounded;
-		bool mbForwardBlocked;
-
-		/// <summary> 점프 상태가 시작될때 발생하는 이벤트 </summary>
-		std::function<void()> mJumpEvent;
-
-		/// <summary> 점프가 끝나고 착지한 상태가 될 때 발생하는 이벤트 </summary>
-		std::function<void()> mGroundEvent;
 
 	private:
 		float mJumpTimer;
